HAAR FINISHED

+++ integral_function/int_image.h
@@ -2,7 +2,7 @@
 # define _INTEGRAL_FUNCTION_H
 
 #include <stdlib.h>
-unsigned long sumImagePart(unsigned long** integralImage,int x1,int y1,int x2, int y2);
+unsigned long sumImagePart(unsigned long** integralImage,unsigned long x1,unsigned long y1,unsigned long x2, unsigned long y2);
 
 
 # endif
